requirements.txt
streamlit==1.28.0
streamlit-authenticator==0.2.3
streamlit_option_menu==0.3.6
python-dotenv==1.0.0
cx_Oracle==8.3.0